fix: creates separate match folder
