+++ src/utils/index.ts
@@ -2,11 +2,6 @@ export * from "./generatePolicy";
 export { default as signToken } from "./signToken";
 export { default as getErrorMessage } from "./getErrorMessage";
 export { default as isJsonString } from "./isJsonString";
-export * from "./match/createMatch";
-export * from "./match/matchLogic";
 export { default as normalizeCity } from "./normalizeCity";
 export * from "./validations";
-export * from "./match/createMatch";
-export * from "./match/matchLogic";
-export * from "./match/publicService";
 export * from "./stringfyBigInt";
